# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlianHub API",
    "version": "v2",
    "description": "Paths, methods, auth and tags only. The full reference is docs/API.md. Generated by npm run api:doc; do not edit by hand.",
    "x-built-from": "836a8cf4392e"
  },
...
  "paths": {
...
    "/api/v2/assignment-rules/project/{projectId}": {
      "get": {
        "summary": "Needs a project the caller can open. Answers with `data`: `rules` (`entries` of person and sentence, `fallbackUserId`, `mode`, `onCreate`, `onChange`) and `ai.state`, whether AI is available.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "read"
      },
      "put": {
        "summary": "Needs the permission to change the project's details. Reads `entries` (a list of at most 50 objects with `userId` and `when`, a sentence of at most 300 characters, one per person), `fallbackUserId` (id or empty), `mode` (`suggest` or `apply`, `suggest` when left out), `onCreate` (boolean, true when left out) and `onChange` (boolean, false when left out) from the body; each person must hold a seat and be able to open the project. Answers with `data`: the saved rules.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/assignment-rules/project/{projectId}/draft": {
      "post": {
        "summary": "Needs the permission to change the project's details, and AI must be on with a model configured (409 otherwise). Reads `userIds` (a list of 1 to 10 ids) from the body. Answers with `data.drafts`: a suggested rule sentence for each person who can open the project, based on their recent tasks there.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/assignment-rules/task/{taskId}": {
      "get": {
        "summary": "Needs a task the caller can open (404 otherwise). Answers with `data`: `decision`, the latest open suggestion for the task or `null`, and `pending`, whether one is still being worked out.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "read"
      }
    },
    "/api/v2/assignment-rules/task/{taskId}/decisions/{decisionId}/accept": {
      "post": {
        "summary": "Needs the permission to change a task's assignee in the task's project. Assigns the suggested person to the task; 409 when the suggestion was already handled, the task already has an assignee, or the person can no longer open it. Answers with `data`: the decision.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "decisionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/assignment-rules/task/{taskId}/decisions/{decisionId}/dismiss": {
      "post": {
        "summary": "Needs the permission to change a task's assignee in the task's project. Marks the suggestion dismissed. Answers with `data`: the decision.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "decisionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/assignment-rules/task/{taskId}/decisions/{decisionId}/undo": {
      "post": {
        "summary": "Needs the permission to change a task's assignee in the task's project. Takes back an accepted suggestion. Answers with `data`: the decision.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "decisionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal access token."
      },
      "companyId": {
        "type": "apiKey",
        "in": "header",
        "name": "companyid",
        "description": "The workspace the request is for."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "The web app's signed-in session."
      }
    }
  }
}